Change the domain timer frequency type
Fixes libvirt commit: bd154804b2ac25c430271dffc4c53cbc0076bfa0

+++ domain.go
@@ -2056,7 +2056,7 @@ type DomainTimer struct {
 	Track      string              `xml:"track,attr,omitempty"`
 	TickPolicy string              `xml:"tickpolicy,attr,omitempty"`
 	CatchUp    *DomainTimerCatchUp `xml:"catchup"`
-	Frequency  uint32              `xml:"frequency,attr,omitempty"`
+	Frequency  uint64              `xml:"frequency,attr,omitempty"`
 	Mode       string              `xml:"mode,attr,omitempty"`
 	Present    string              `xml:"present,attr,omitempty"`
 }
